Code updates for the new analyzer. (#22)
- Requiring CMake 3.6.
- Removed obsolete configs and code.
- sources.cfg: updated to follow new configuration structure.
- AnalysisCategories.h:
  - renamed DataCategoryType to SampleType;
  - reimplemented EventRegion, EventCategory and EventSubCategory.
- BaseEventAnalyzer.h:
  - major code redesign;
  - histograms, tables and datacard shapes production code moved into separate files.
- EventAnalyzerData.h: fixed to be compilable + reviewed specialisation implementation.
- EventAnalyzerDataCollection.h: reimplemented EventAnalyzerDataId and collection interface.
- SampleDescriptor.h: updated configuration structure.
- bb*tauAnalyzer: modified to be compatible with the new BaseEventAnalyzer.
- Lester_mt2_bisect.h moved into Instruments/include
- SampleDescriptor.h and TTFileMerger.cxx: fixed compilation warnings.
- NJets_HT_weight: made GetWeight public for test purposes.

+++ Instruments/include/SampleDescriptor.h
@@ -35,7 +35,7 @@ struct SampleDescriptor {
     {
         size_t total_n_events = 0;
         for (const auto& genEventCount : gen_counts){
-            const size_t nevents = genEventCount.second;
+            const size_t nevents = static_cast<size_t>(genEventCount.second);
             total_n_events += nevents;
         }
         return total_n_events;
